camera_streamer/requirements.txt
# gRPC 依赖
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0

# 图像处理
opencv-python>=4.5.0
numpy>=1.20.0
